+++ dash_bio/Ideogram.py
@@ -4,7 +4,7 @@ from dash.development.base_component import Component, _explicitize_args
 
 
 class Ideogram(Component):
-    """A Ideogram component.
+    """An Ideogram component.
 The Ideogram component is used to draw and animate genome-wide
 datasets for organisms such as human, mouse, and any other
 eukaryote. The Ideogram component can be used to compare
@@ -20,7 +20,7 @@ Keyword arguments:
 instances.
 - style (dict; optional): The component's inline styles
 - className (string; optional): The CSS class of the component wrapper
-- annotationsLayout (a value equal to: 'tracks', 'histogram', 'overlay'; optional): Layout of ideogram annotations.
+- annotationsLayout (a value equal to: 'tracks', 'histogram', 'overlay'; default 'tracks'): Layout of ideogram annotations.
 One of "tracks", "histogram", or "overlay".
 
 "tracks": display annotations in tracks beside each chromosome.
@@ -30,25 +30,30 @@ cluster/bin is shown as a bar, the height of which represents the number of anno
 genomic range.
 
 "overlay": display annotations directly over chromosomes.
-- annotations (list; optional): A list of annotation objects. Annotation objects can also have a name, color, shape, and
-track index. At the moment there is more keys specified and the docs need updating.
+- annotations (dict; optional): A list of annotation objects. Annotation objects can also have a name, color, shape, and
+track index. At the moment there is more keys specified and the docs need updating. annotations has the following type: list of dicts containing keys 'name', 'chr', 'start', 'stop'.
+Those keys have the following types:
+  - name (string; optional)
+  - chr (string; optional)
+  - start (number; optional)
+  - stop (number; optional)
 - annotationsPath (string; optional): An absolute or relative URL directing to a JSON file containing annotation objects (JSON).
 - annotationsData (string; optional): Use this prop in a dash callback to return annotationData when hovered.
 It is read-only, i.e., it cannot be used with dash.dependencies.Output but only with
 dash.dependencies.Input
-- annotationTracks (list; optional): A list of objects with metadata for each track, e.g., id, display name, color, shape.
+- annotationTracks (list of dicts; optional): A list of objects with metadata for each track, e.g., id, display name, color, shape.
 - annotationHeight (number; optional): Not used if annotationsLayout is set to "overlay".
 The height of histogram bars or the size of annotations tracks symbols
-- annotationsColor (string; optional): Color of annotations.
+- annotationsColor (string; default '#F00'): Color of annotations.
 - histogramScaling (a value equal to: 'absolute', 'relative'; optional): Scaling of histogram bars height
 Only used if annotationsLayout is set to "histogram".
 One of "absolute" or "relative".
 
 "absolute": sets bar height relative to tallest bar in all chromosomes.
 "relative": sets bar height relative to tallest bar in each chromosome.
-- barWidth (number; optional): Pixel width of histogram bars.
+- barWidth (number; default 3): Pixel width of histogram bars.
 Only used if annotationsLayout is set to "histogram".
-- showAnnotTooltip (boolean; optional): Whether to show a tooltip upon mousing over an annotation.
+- showAnnotTooltip (boolean; default True): Whether to show a tooltip upon mousing over an annotation.
 - assembly (string; optional): Default: latest RefSeq assembly for specified organism.
 The genome assembly to display.
 Takes assembly name (e.g., "GRCh37"),
@@ -58,11 +63,9 @@ or GenBank accession (e.g., "GCA_000005005.5")
 chromosome. Useful when ideogram consists of one chromosome and you want to be
 able to focus on a region within that chromosome,
 and create an interactive sliding window to other regions
-- brushData (optional): A dash callback that is activated when the 'brush' prop is used in component.
+- brushData (dict; optional): A dash callback that is activated when the 'brush' prop is used.
 It will return an dictionary like so:
-
 {'start': <value>, 'end': <value>, 'extent': <value>}
-
 where start is the left most edge, end is right most edge, and extent is the total width of
 the brush.
 It is read-only, i.e., it cannot be used with dash.dependencies.Output but only with
@@ -73,10 +76,10 @@ Those keys have the following types:
   - extent (string; optional)
 - container (string; optional): CSS styling and the id of the container holding the Ideogram in
 react-ideogram.js, this is where all the d3 magic happens.
-- chrHeight (number; optional): The pixel height of the tallest chromosome in the ideogram
-- chrMargin (number; optional): The pixel space of margin between each chromosome.
-- chrWidth (number; optional): The pixel width of each chromosome.
-- chromosomes (list | dict; optional): A list of the names of chromosomes to display. Useful for depicting a subset of the
+- chrHeight (number; default 400): The pixel height of the tallest chromosome in the ideogram
+- chrMargin (number; default 10): The pixel space of margin between each chromosome.
+- chrWidth (number; default 10): The pixel width of each chromosome.
+- chromosomes (list of strings | dict; optional): A list of the names of chromosomes to display. Useful for depicting a subset of the
 chromosomes in the genome, e.g., a single chromosome.
 
 If Homology (between two different species):
@@ -87,15 +90,15 @@ Ex: chromosomes={
 
 General case to specify specific chromosomes:
 Ex: chromosomes=['1', '2']
-- dataDir (string; optional): Absolute or relative URL of the directory containing data needed to draw banded chromosomes.
+- dataDir (string; default 'https://unpkg.com/ideogram@1.5.0/dist/data/bands/native/'): Absolute or relative URL of the directory containing data needed to draw banded chromosomes.
 You will need to set up your own database to grab data from a custom database.
-- organism (string | number; optional): Organism(s) to show chromosomes for. Supply organism's name as a string (e.g., "human") or
+- organism (string | number; default 'human'): Organism(s) to show chromosomes for. Supply organism's name as a string (e.g., "human") or
 organism's NCBI Taxonomy ID (taxid, e.g., 9606) to display chromosomes from a single
 organism, or an array of organisms' names or taxids to display chromosomes from multiple
 species.
 - localOrganism (dict; optional): Provide local JSON organism into this prop from a local user JSON file.
 DataDir must not be initialized.
-- homology (optional): Used to compare two chromosomes with each other.
+- homology (dict; optional): Used to compare two chromosomes.
 The keys "chrOne" and "chrTwo" represent one chromosome each. Organism is the taxID or name.
 Start is an array, containing start one and start two, in this order. Stop is an array,
 containing stop one, and stop two, in this order.
@@ -112,16 +115,16 @@ Ex: homology={
     }
 }. homology has the following type: dict containing keys 'chrOne', 'chrTwo'.
 Those keys have the following types:
-  - chrOne (optional): . chrOne has the following type: dict containing keys 'organism', 'start', 'stop'.
+  - chrOne (dict; optional): chrOne has the following type: dict containing keys 'organism', 'start', 'stop'.
 Those keys have the following types:
   - organism (string; required)
-  - start (list; optional)
-  - stop (list; optional)
-  - chrTwo (optional): . chrTwo has the following type: dict containing keys 'organism', 'start', 'stop'.
+  - start (list of numbers; optional)
+  - stop (list of numbers; optional)
+  - chrTwo (dict; optional): chrTwo has the following type: dict containing keys 'organism', 'start', 'stop'.
 Those keys have the following types:
   - organism (string; required)
-  - start (list; optional)
-  - stop (list; optional)
+  - start (list of numbers; optional)
+  - stop (list of numbers; optional)
 - perspective (a value equal to: 'comparative'; optional): Use perspective: 'comparative' to enable annotations between two chromosomes,
 either within the same organism or different organisms. Used for homology.
 - fullChromosomeLabels (boolean; optional): Whether to include abbreviation species name in chromosome label. Used for homology.
@@ -130,21 +133,21 @@ The quantity refers to an approximate value in bands per haploid set (bphs).
 One of 450, 550, or 850.
 - filterable (boolean; optional): Whether annotations should be filterable or not.
 - orientation (a value equal to: 'vertical', 'horizontal'; optional): The orientation of chromosomes on the page.
-- ploidy (number; optional): The ploidy - number of chromosomes to depict for each chromosome set.
-- ploidyDesc (list; optional): Description of ploidy in each chromosome set in terms of ancestry composition.
+- ploidy (number; default 1): The ploidy - number of chromosomes to depict for each chromosome set.
+- ploidyDesc (list of dicts; optional): Description of ploidy in each chromosome set in terms of ancestry composition.
 - ancestors (dict; optional): A map associating ancestor labels to colors. Used to color
 chromosomes from different ancestors in polyploid genomes.
-- rangeSet (list; optional): List of objects describing segments of recombination among chromosomes in a chromosome set.
-- rotatable (boolean; optional): Whether chromosomes are rotatable on click.
+- rangeSet (list of dicts; optional): List of objects describing segments of recombination among chromosomes in a chromosome set.
+- rotatable (boolean; default True): Whether chromosomes are rotatable on click.
 - rotated (boolean; optional): Dash callback that returns true if rotated, and false if not.
 - sex (a value equal to: 'male', 'female'; optional): Useful for omitting chromosome Y in female animals.
 Currently only supported for organisms that use XY sex-determination.
-- showChromosomeLabels (boolean; optional): Whether to show chromosome labels, e.g., 1, 2, 3, X, Y.
-- showBandLabels (boolean; optional): Whether to show cytogenetic band labels, e.g., 1q21
-- showFullyBanded (boolean; optional): Whether to show fully banded chromosomes for genomes that have sufficient data. Useful for
+- showChromosomeLabels (boolean; default True): Whether to show chromosome labels, e.g., 1, 2, 3, X, Y.
+- showBandLabels (boolean; default False): Whether to show cytogenetic band labels, e.g., 1q21.
+- showFullyBanded (boolean; default True): Whether to show fully banded chromosomes for genomes that have sufficient data. Useful for
 showing simpler chromosomes of cytogenetically well-characterized organisms, e.g., human,
 beside chromosomes of less studied organisms, e.g., chimpanzee.
-- showNonNuclearChromosomes (boolean; optional): Whether to show non-nuclear chromosomes,
+- showNonNuclearChromosomes (boolean; default False): Whether to show non-nuclear chromosomes,
 e.g., for mitochondrial (MT) and chloroplast (CP) DNA."""
     @_explicitize_args
     def __init__(self, id=Component.REQUIRED, style=Component.UNDEFINED, className=Component.UNDEFINED, annotationsLayout=Component.UNDEFINED, annotations=Component.UNDEFINED, annotationsPath=Component.UNDEFINED, annotationsData=Component.UNDEFINED, annotationTracks=Component.UNDEFINED, annotationHeight=Component.UNDEFINED, annotationsColor=Component.UNDEFINED, histogramScaling=Component.UNDEFINED, barWidth=Component.UNDEFINED, showAnnotTooltip=Component.UNDEFINED, assembly=Component.UNDEFINED, brush=Component.UNDEFINED, brushData=Component.UNDEFINED, container=Component.UNDEFINED, chrHeight=Component.UNDEFINED, chrMargin=Component.UNDEFINED, chrWidth=Component.UNDEFINED, chromosomes=Component.UNDEFINED, dataDir=Component.UNDEFINED, organism=Component.UNDEFINED, localOrganism=Component.UNDEFINED, homology=Component.UNDEFINED, perspective=Component.UNDEFINED, fullChromosomeLabels=Component.UNDEFINED, resolution=Component.UNDEFINED, filterable=Component.UNDEFINED, orientation=Component.UNDEFINED, ploidy=Component.UNDEFINED, ploidyDesc=Component.UNDEFINED, ancestors=Component.UNDEFINED, rangeSet=Component.UNDEFINED, rotatable=Component.UNDEFINED, rotated=Component.UNDEFINED, sex=Component.UNDEFINED, showChromosomeLabels=Component.UNDEFINED, showBandLabels=Component.UNDEFINED, showFullyBanded=Component.UNDEFINED, showNonNuclearChromosomes=Component.UNDEFINED, **kwargs):