+++ tests/unit/SequenceViewer.test.js
@@ -1,0 +1,10 @@
+import SequenceViewer from '../../src/lib/components/SequenceViewer.react.js';
+import React from 'react';
+import {mount, render} from 'enzyme';
+import seq from './sequence_viewer_data.json';
+
+
+test('SequenceViewer renders', () => {
+    const component = render(<SequenceViewer sequence={seq.sequence} />);
+    expect(component.html()).toBeDefined();
+});